// FE/medical-supplies-website/src/app/module/employee/model/EmployeeUserDetail.ts
export interface EmployeeUserDetail {
  status: number;
  employeeId?: number;
  employeeCode?: string;
  employeeName?: string;
  phone?: string;
  gender?: boolean;
  employeeAddress?: string;
  dateOfBirth?: string;
  idCard?: string;
  salary?: number;
  employeeImg?: string;
  username?: string;
  accountEmail?: string;
  positionName?: string;
}